Updated modules to new structure, added full path to quay.io biocontainers, modified VarDictJava module to output *.vcf.gz files to avoid clash with bedtools intersect output.
