+++ conf/test.config
@@ -25,4 +25,5 @@ params {
     input  = "https://raw.githubusercontent.com/CIBERER-pipelines/mosaicism_nextflow/main/testdata/samplesheet_test.csv"
     fasta = "https://raw.githubusercontent.com/CIBERER-pipelines/mosaicism_nextflow/main/testdata/chr17_var_regions.fa"
     fai = "https://raw.githubusercontent.com/CIBERER-pipelines/mosaicism_nextflow/main/testdata/chr17_var_regions.fa.fai"
+    chrom_sizes = "https://raw.githubusercontent.com/CIBERER-pipelines/mosaicism_nextflow/main/testdata/chr17_chrom_sizes.txt"
 }
