+++ modules/nf-core/bedtools/intersect/environment.yml
@@ -1,0 +1,7 @@
+name: bedtools_intersect
+channels:
+  - conda-forge
+  - bioconda
+  - defaults
+dependencies:
+  - bioconda::bedtools=2.31.1
