+++ modules/nf-core/bedtools/intersect/meta.yml
@@ -1,0 +1,59 @@
+name: bedtools_intersect
+description: Allows one to screen for overlaps between two sets of genomic features.
+keywords:
+  - bed
+  - intersect
+  - overlap
+tools:
+  - bedtools:
+      description: |
+        A set of tools for genomic analysis tasks, specifically enabling genome arithmetic (merge, count, complement) on various file types.
+      documentation: https://bedtools.readthedocs.io/en/latest/content/tools/intersect.html
+      licence: ["MIT"]
+input:
+  - meta:
+      type: map
+      description: |
+        Groovy Map containing sample information
+        e.g. [ id:'test', single_end:false ]
+  - intervals1:
+      type: file
+      description: BAM/BED/GFF/VCF
+      pattern: "*.{bam|bed|gff|vcf}"
+  - intervals2:
+      type: file
+      description: BAM/BED/GFF/VCF
+      pattern: "*.{bam|bed|gff|vcf}"
+  - meta2:
+      type: map
+      description: |
+        Groovy Map containing reference chromosome sizes
+        e.g. [ id:'test' ]
+  - chrom_sizes:
+      type: file
+      description: Chromosome sizes file
+      pattern: "*{.sizes,.txt}"
+output:
+  - meta:
+      type: map
+      description: |
+        Groovy Map containing sample information
+        e.g. [ id:'test', single_end:false ]
+  - intersect:
+      type: file
+      description: File containing the description of overlaps found between the two features
+      pattern: "*.${extension}"
+  - versions:
+      type: file
+      description: File containing software versions
+      pattern: "versions.yml"
+authors:
+  - "@edmundmiller"
+  - "@sruthipsuresh"
+  - "@drpatelh"
+  - "@sidorov-si"
+maintainers:
+  - "@edmundmiller"
+  - "@sruthipsuresh"
+  - "@drpatelh"
+  - "@sidorov-si"
