+++ modules/nf-core/custom/dumpsoftwareversions/environment.yml
@@ -1,0 +1,7 @@
+name: custom_dumpsoftwareversions
+channels:
+  - conda-forge
+  - bioconda
+  - defaults
+dependencies:
+  - bioconda::multiqc=1.20
