--- modules/nf-core/modules/custom/dumpsoftwareversions/meta.yml
@@ -1,34 +1,0 @@
-name: custom_dumpsoftwareversions
-description: Custom module used to dump software versions within the nf-core pipeline template
-keywords:
-  - custom
-  - version
-tools:
-  - custom:
-      description: Custom module used to dump software versions within the nf-core pipeline template
-      homepage: https://github.com/nf-core/tools
-      documentation: https://github.com/nf-core/tools
-      licence: ["MIT"]
-input:
-  - versions:
-      type: file
-      description: YML file containing software versions
-      pattern: "*.yml"
-
-output:
-  - yml:
-      type: file
-      description: Standard YML file containing software versions
-      pattern: "software_versions.yml"
-  - mqc_yml:
-      type: file
-      description: MultiQC custom content YML file containing software versions
-      pattern: "software_versions_mqc.yml"
-  - versions:
-      type: file
-      description: File containing software versions
-      pattern: "versions.yml"
-
-authors:
-  - "@drpatelh"
-  - "@grst"
