--- modules/nf-core/modules/vardictjava/meta.yml
@@ -1,60 +1,0 @@
-name: "vardictjava"
-
-description: The Java port of the VarDict variant caller
-keywords:
-  - variant calling
-  - VarDict
-  - AstraZeneca
-tools:
-  - "vardictjava":
-      description: "Java port of the VarDict variant discovery program"
-      homepage: "https://github.com/AstraZeneca-NGS/VarDictJava"
-      documentation: "https://github.com/AstraZeneca-NGS/VarDictJava"
-      tool_dev_url: "https://github.com/AstraZeneca-NGS/VarDictJava"
-      doi: "10.1093/nar/gkw227 "
-      licence: "['MIT']"
-
-input:
-  - meta:
-      type: map
-      description: |
-        Groovy Map containing sample information
-        e.g. [ id:'test', single_end:false ]
-  - bam:
-      type: file
-      description: BAM/SAM file
-      pattern: "*.{bam,sam}"
-  - bai:
-      type: file
-      description: Index of the BAM file
-      pattern: "*.bai"
-  - bed:
-      type: file
-      description: BED with the regions of interest
-      pattern: "*.bed"
-  - fasta:
-      type: file
-      description: FASTA of the reference genome
-      pattern: "*.{fa,fasta}"
-  - fasta_fai:
-      type: file
-      description: The index of the FASTA of the reference genome
-      pattern: "*.fai"
-
-output:
-  - meta:
-      type: map
-      description: |
-        Groovy Map containing sample information
-        e.g. [ id:'test', single_end:false ]
-  - versions:
-      type: file
-      description: File containing software versions
-      pattern: "versions.yml"
-  - vcf:
-      type: file
-      description: VCF file output
-      pattern: "*.vcf.gz"
-
-authors:
-  - "@nvnieuwk"
