+++ modules/nf-core/samtools/mpileup/environment.yml
@@ -1,0 +1,8 @@
+name: samtools_mpileup
+channels:
+  - conda-forge
+  - bioconda
+  - defaults
+dependencies:
+  - bioconda::samtools=1.19.2
+  - bioconda::htslib=1.19.1
