+++ modules/nf-core/samtools/mpileup/tests/tags.yml
@@ -1,0 +1,2 @@
+samtools/mpileup:
+  - "modules/nf-core/samtools/mpileup/**"
