+++ modules/nf-core/samtools/sort/tests/nextflow.config
@@ -1,0 +1,7 @@
+process {
+
+    withName: SAMTOOLS_SORT {
+        ext.prefix = { "${meta.id}.sorted" }
+    }
+
+}
