+++ modules/nf-core/samtools/sort/tests/tags.yml
@@ -1,0 +1,3 @@
+samtools/sort:
+  - modules/nf-core/samtools/sort/**
+  - tests/modules/nf-core/samtools/sort/**
