+++ modules/nf-core/tabix/bgzip/environment.yml
@@ -1,0 +1,7 @@
+name: tabix_bgzip
+channels:
+  - conda-forge
+  - bioconda
+  - defaults
+dependencies:
+  - bioconda::tabix=1.11
