+++ modules/nf-core/tabix/bgzip/meta.yml
@@ -1,0 +1,52 @@
+name: tabix_bgzip
+description: Compresses/decompresses files
+keywords:
+  - compress
+  - decompress
+  - bgzip
+  - tabix
+tools:
+  - bgzip:
+      description: |
+        Bgzip compresses or decompresses files in a similar manner to, and compatible with, gzip.
+      homepage: https://www.htslib.org/doc/tabix.html
+      documentation: http://www.htslib.org/doc/bgzip.html
+      doi: 10.1093/bioinformatics/btp352
+      licence: ["MIT"]
+input:
+  - meta:
+      type: map
+      description: |
+        Groovy Map containing sample information
+        e.g. [ id:'test', single_end:false ]
+  - input:
+      type: file
+      description: file to compress or to decompress
+output:
+  - meta:
+      type: map
+      description: |
+        Groovy Map containing sample information
+        e.g. [ id:'test', single_end:false ]
+  - output:
+      type: file
+      description: Output compressed/decompressed file
+      pattern: "*."
+  - gzi:
+      type: file
+      description: Optional gzip index file for compressed inputs
+      pattern: "*.gzi"
+  - versions:
+      type: file
+      description: File containing software versions
+      pattern: "versions.yml"
+authors:
+  - "@joseespinosa"
+  - "@drpatelh"
+  - "@maxulysse"
+  - "@nvnieuwk"
+maintainers:
+  - "@joseespinosa"
+  - "@drpatelh"
+  - "@maxulysse"
+  - "@nvnieuwk"
