+++ modules/nf-core/vardictjava/meta.yml
@@ -1,0 +1,70 @@
+name: "vardictjava"
+description: The Java port of the VarDict variant caller
+keywords:
+  - variant calling
+  - vcf
+  - bam
+  - snv
+  - sv
+tools:
+  - "vardictjava":
+      description: "Java port of the VarDict variant discovery program"
+      homepage: "https://github.com/AstraZeneca-NGS/VarDictJava"
+      documentation: "https://github.com/AstraZeneca-NGS/VarDictJava"
+      tool_dev_url: "https://github.com/AstraZeneca-NGS/VarDictJava"
+      doi: "10.1093/nar/gkw227 "
+      licence: "['MIT']"
+input:
+  - meta:
+      type: map
+      description: |
+        Groovy Map containing sample information
+        e.g. [ id:'test', single_end:false ]
+  - bams:
+      type: file
+      description: One or two BAM files. Supply two BAM files to run Vardict in paired mode.
+      pattern: "*.bam"
+  - bais:
+      type: file
+      description: Index/indices of the BAM file(s)
+      pattern: "*.bai"
+  - bed:
+      type: file
+      description: BED with the regions of interest
+      pattern: "*.bed"
+  - meta2:
+      type: map
+      description: |
+        Groovy Map containing fasta information
+        e.g. [ id:'test', single_end:false ]
+  - fasta:
+      type: file
+      description: FASTA of the reference genome
+      pattern: "*.{fa,fasta}"
+  - meta3:
+      type: map
+      description: |
+        Groovy Map containing fasta information
+        e.g. [ id:'test', single_end:false ]
+  - fasta_fai:
+      type: file
+      description: The index of the FASTA of the reference genome
+      pattern: "*.fai"
+output:
+  - meta:
+      type: map
+      description: |
+        Groovy Map containing sample information
+        e.g. [ id:'test', single_end:false ]
+  - versions:
+      type: file
+      description: File containing software versions
+      pattern: "versions.yml"
+  - vcf:
+      type: file
+      description: VCF file output
+      pattern: "*.vcf.gz"
+authors:
+  - "@nvnieuwk"
+maintainers:
+  - "@nvnieuwk"
