+++ modules/nf-core/vardictjava/tests/tags.yml
@@ -1,0 +1,2 @@
+vardictjava:
+  - modules/nf-core/vardictjava/**
