+++ workflows/mosaicism.nf
@@ -18,6 +18,7 @@ for (param in checkPathParamList) { if (param) { file(param, checkIfExists: true
 if (params.input) { ch_input = file(params.input) } else { exit 1, 'Input samplesheet not specified!' }
 if (params.fasta) { ch_fasta = file(params.fasta) } else { exit 1, 'Fasta file not specified!' }
 if (params.fai) { ch_fasta_fai = file(params.fai) } else { exit 1, 'Fai file not specified!' }
+if (params.chrom_sizes) { ch_chrom_sizes = file(params.chrom_sizes) } else { exit 1, 'chrom.sizes file not specified!' }
 
 /*
 ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
@@ -52,9 +53,9 @@ include { VARSCAN_WF } from '../subworkflows/local/varscan_workflow'
 // MODULE: Installed directly from nf-core/modules
 //
 
-include { CUSTOM_DUMPSOFTWAREVERSIONS } from '../modules/nf-core/modules/custom/dumpsoftwareversions/main'
-include { VARDICTJAVA } from '../modules/nf-core/modules/vardictjava/main'
-include { BEDTOOLS_INTERSECT } from '../modules/nf-core/modules/bedtools/intersect/main'
+include { CUSTOM_DUMPSOFTWAREVERSIONS } from '../modules/nf-core/custom/dumpsoftwareversions/main'
+include { VARDICTJAVA } from '../modules/nf-core/vardictjava/main'
+include { BEDTOOLS_INTERSECT } from '../modules/nf-core/bedtools/intersect/main'
 
 /*
 ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
@@ -89,9 +90,17 @@ workflow MOSAICISM {
   //
   // MODULE: Run Vardictjava
   //
+  
+  Channel.fromPath(ch_fasta)
+    .map{ create_fa_meta(it) }
+    .set{ ch_fasta_meta }
+  
+  Channel.fromPath(ch_fasta_fai)
+    .map{ create_fai_meta(it) }
+    .set{ ch_fai_meta }
 
   VARDICTJAVA (
-    INPUT_CHECK.out.reads_bam_bai_bed, ch_fasta, ch_fasta_fai
+    INPUT_CHECK.out.reads_bam_bai_bed, ch_fasta_meta, ch_fai_meta
   )
 
   ch_versions = ch_versions.mix(VARDICTJAVA.out.versions.first())
@@ -100,9 +109,13 @@ workflow MOSAICISM {
   //ch_versions.mix(VARDICTJAVA.out.versions.view())
 
   ch_bedtools = VARDICTJAVA.out.vcf.join(VARSCAN_WF.out.varscan_out)
+  
+  Channel.fromPath(ch_chrom_sizes)
+    .map{ create_chrom_sizes_input(it) }
+    .set{ ch_chrom_sizes_meta }
 
   BEDTOOLS_INTERSECT (
-    ch_bedtools, "vcf"
+    ch_bedtools, ch_chrom_sizes_meta
   )
 
   ch_versions = ch_versions.mix(BEDTOOLS_INTERSECT.out.versions.first())
@@ -110,6 +123,63 @@ workflow MOSAICISM {
   CUSTOM_DUMPSOFTWAREVERSIONS (
       ch_versions.unique().collectFile(name: 'collated_versions.yml')
   )
+
+}
+
+// Add meta for fasta input for VARDICTJAVA
+
+def create_fa_meta (fasta) {
+    // create meta map
+    def meta = [:]
+    meta.id = fasta.baseName
+    
+    // add path of fasta file to the meta map
+    def chrom_sizes_meta = []
+    if (!file(fasta).exists()) {
+        exit 1, "ERROR: Please check input parameters -> fasta file does not exist!\n${fasta}"
+    }
+
+    fasta_meta = [ meta, file(fasta) ]
+
+    return fasta_meta
+
+}
+
+// Add meta for fai input for VARDICTJAVA
+
+def create_fai_meta (fai) {
+    // create meta map
+    def meta = [:]
+    meta.id = fai.baseName
+    
+    // add path of fasta file to the meta map
+    def fai_meta = []
+    if (!file(fai).exists()) {
+        exit 1, "ERROR: Please check input parameters -> fai file does not exist!\n${fai}"
+    }
+
+    fai_meta = [ meta, file(fai) ]
+
+    return fai_meta
+
+}
+
+// Add meta for chrom.sizes input for BEDTOOLS_INTERSECT
+
+def create_chrom_sizes_input (chrom_sizes) {
+    // create meta map
+    def meta = [:]
+    meta.id = chrom_sizes.baseName
+    
+    // add path of the chrom.sizes file to the meta map
+    def chrom_sizes_meta = []
+    if (!file(chrom_sizes).exists()) {
+        exit 1, "ERROR: Please check input parameters -> chrom_sizes file does not exist!\n${chrom_sizes}"
+    }
+
+    chrom_sizes_meta = [ meta, file(chrom_sizes) ]
+
+    return chrom_sizes_meta
 
 }
